v2.4.0.11 -> v2.4.0.12
  - Get non-cpuid Cyrix probing right (it's not a NexGen)
  - Jens Axboe: cdrom tray status and queing cleanups
  - AGP GART: don't disable VIA, and allow i815 with external AGP
  - Coda: use iget4() in order to have big inode numbers without clashes.
  - Fix UDF writepage() page locking
  - NIIBE Yutaka: SuperH update
  - Martin Diehl and others: SiS pirq routing fixes
  - Andy Grover: ACPI update
  - Andrea Arkangeli: LVM update
  - Ingo Molnar: RAID cleanups
  - David Miller: sparc and networking updates
  - Make NFS really be able to handle large files

+++ include/linux/nfs_xdr.h
@@ -74,7 +74,7 @@ struct nfs_readargs {
 
 struct nfs_readres {
 	struct nfs_fattr *	fattr;
-	unsigned int		count;
+	__u32			count;
 	int                     eof;
 };
 
@@ -84,7 +84,7 @@ struct nfs_readres {
 #define NFS_WRITE_MAXIOV        8
 struct nfs_writeargs {
 	struct nfs_fh *		fh;
-	__u32			offset;
+	__u64			offset;
 	__u32			count;
 	enum nfs3_stable_how	stable;
 	unsigned int		nriov;
